+++ src/lab1/string.h
@@ -3,12 +3,12 @@
 #define _STRING_H
 
 int strcmp(const char* s1, const char* s2);
-int strncmp(const char* s1, const char* s2, int len);
+int strncmp(const char* s1, const char* s2, size_t len);
 uint strlen(const char* s);
 uint isspace(char c);
 void trim(char* str);
-void memset(char* s, char value, uint size);
-void strins(char* s, char c, int pos);
-void strdel(char* s, int pos);
+void memset(char* s, char value, size_t size);
+void strins(char* s, char c, size_t pos);
+void strdel(char* s, size_t pos);
 
 #endif
